Update bayes-algorithm.py
added a column to results to track whether a disease has any positive or negative manifestations. If a disease is never modified by a manifestation, i get rid of it from  the results data frame prior to calculating conditional probabilities.

Also fixed an error in calculation of the fdr and specificity.  False discover rate is 1 - ppv.